+++ spec/series_spec.rb
@@ -122,4 +122,12 @@ describe Rubicure::Series do
     end
   end
 
+  describe "#each_with_girls" do
+    subject{ series.each_with_girls }
+
+    let(:series)     { Rubicure::Series.find(series_name) }
+    let(:series_name){ :splash_star }
+
+    it { expect { |b| series.each_with_girls(&b) }.to yield_successive_args(Rubicure::Girl, Rubicure::Girl) }
+  end
 end